+++ app/services/flow/flows/provider_capital.rb
@@ -93,7 +93,7 @@ module Flow
           end,
         },
         employment_incomes: {
-          path: ->(application) { urls.providers_legal_aid_application_employment_income_path(application) },
+          path: ->(application) { urls.providers_legal_aid_application_means_employment_income_path(application) },
           forward: ->(application) { application.income_types? ? :income_summary : :no_income_summaries },
         },
         full_employment_details: {
