+++ app/views/providers/means/employment_incomes/show.html.erb
@@ -1,5 +1,5 @@
 <%= form_with(model: @form,
-              url: providers_legal_aid_application_employment_income_path(@legal_aid_application),
+              url: providers_legal_aid_application_means_employment_income_path(@legal_aid_application),
               method: :patch,
               local: true) do |form| %>
