+++ app/views/shared/check_answers/_employed_income.html.erb
@@ -5,7 +5,7 @@
   <div class="govuk-grid-column-one-third govuk-summary-list--no-border align-text-right">
     <% unless read_only %>
       <p>
-        <%= link_to_accessible(t('generic.change'), providers_legal_aid_application_employment_income_path(@legal_aid_application),
+        <%= link_to_accessible(t('generic.change'), providers_legal_aid_application_means_employment_income_path(@legal_aid_application),
                              class: 'govuk-link change-link') %>
       </p>
     <% end %>
