+++ config/locales/en/providers.yml
@@ -437,22 +437,6 @@ en:
       show:
         h1-heading: Details of the latest domestic abuse incident
         details_label: What happened during the incident?
-    employment_incomes:
-      show:
-        hmrc-information: The information on this page has been provided by HMRC.
-        employed: Review %{name}'s employment income
-        not_employed: HMRC found a record of your client's employment
-        hmrc_not_employed: "You told us your client is not employed but HMRC shared the following information with us:"
-        supplementary_question: Do you need to tell us anything else about your client's employment?
-        employment_hint: For example, if they get cash in hand pay, have zero-hours contracts, or recently started or ended employment.
-        date: Date
-        income: Income
-        deductions: Deductions
-        gross: Gross
-        enter_details: Enter details
-        details_hint: You'll need to upload supporting evidence later.
-        tax: Tax
-        ni: National Insurance
     end_of_applications:
       show:
         heading: Application complete
@@ -647,6 +631,22 @@ en:
           details_4: any benefits in kind they received
           employment_details_header: Enter your client's employment details
           employment_details_hint: You'll need to upload supporting evidence later.
+      employment_incomes:
+        show:
+          hmrc-information: The information on this page has been provided by HMRC.
+          employed: Review %{name}'s employment income
+          not_employed: HMRC found a record of your client's employment
+          hmrc_not_employed: "You told us your client is not employed but HMRC shared the following information with us:"
+          supplementary_question: Do you need to tell us anything else about your client's employment?
+          employment_hint: For example, if they get cash in hand pay, have zero-hours contracts, or recently started or ended employment.
+          date: Date
+          income: Income
+          deductions: Deductions
+          gross: Gross
+          enter_details: Enter details
+          details_hint: You'll need to upload supporting evidence later.
+          tax: Tax
+          ni: National Insurance
     means_summaries:
       show:
         h1-heading: Check your answers
