+++ spec/requests/providers/client_completed_means_spec.rb
@@ -93,7 +93,7 @@ RSpec.describe Providers::ClientCompletedMeansController, type: :request do
 
             it "redirects to the employed income page" do
               subject
-              expect(response).to redirect_to(providers_legal_aid_application_employment_income_path(legal_aid_application))
+              expect(response).to redirect_to(providers_legal_aid_application_means_employment_income_path(legal_aid_application))
             end
           end
